// src/app/api/fn/inventory/inventory-counts-list.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { HttpClient, HttpContext, HttpResponse } from '@angular/common/http';
import { Observable } from 'rxjs';
import { filter, map } from 'rxjs/operators';
import { StrictHttpResponse } from '../../strict-http-response';
import { RequestBuilder } from '../../request-builder';

import { PaginatedStockCountList } from '../../models/paginated-stock-count-list';

export interface InventoryCountsList$Params {

/**
 * Category id.
 */
  category?: number;

/**
 * On or after this date.
 */
  date_from?: string;

/**
 * On or before this date.
 */
  date_to?: string;

/**
 * A page number within the paginated result set.
 */
  page?: number;

/**
 * Matches codes, names and numbers.
 */
  search?: string;

/**
 * Document status.
 */
  status?: 'DRAFT' | 'POSTED';
}

export function inventoryCountsList(http: HttpClient, rootUrl: string, params?: InventoryCountsList$Params, context?: HttpContext): Observable<StrictHttpResponse<PaginatedStockCountList>> {
  const rb = new RequestBuilder(rootUrl, inventoryCountsList.PATH, 'get');
  if (params) {
    rb.query('category', params.category, {});
    rb.query('date_from', params.date_from, {});
    rb.query('date_to', params.date_to, {});
    rb.query('page', params.page, {});
    rb.query('search', params.search, {});
    rb.query('status', params.status, {});
  }

  return http.request(
    rb.build({ responseType: 'json', accept: 'application/json', context })
  ).pipe(
    filter((r: any): r is HttpResponse<any> => r instanceof HttpResponse),
    map((r: HttpResponse<any>) => {
      return r as StrictHttpResponse<PaginatedStockCountList>;
    })
  );
}

inventoryCountsList.PATH = '/api/inventory/counts/';